Fix build: remove CopilotKit hooks, fix INTEGER overflow in compensation queries

+++ app/fractional-ceo-jobs-uk/page.tsx
@@ -62,7 +62,7 @@ async function getExecutiveStats() {
     const sql = createDbQuery()
     const [totalResult, avgRateResult, remoteResult] = await Promise.all([
       sql`SELECT COUNT(*) as count FROM jobs WHERE is_active = true AND role_category = 'Executive'`,
-      sql`SELECT AVG(CAST(REGEXP_REPLACE(compensation, '[^0-9]', '', 'g') AS INTEGER)) as avg FROM jobs WHERE is_active = true AND role_category = 'Executive' AND compensation IS NOT NULL AND compensation ~ '^[£$]?[0-9]+'`,
+      sql`SELECT AVG(CAST(REGEXP_REPLACE(compensation, '[^0-9]', '', 'g') AS BIGINT)) as avg FROM jobs WHERE is_active = true AND role_category = 'Executive' AND compensation IS NOT NULL AND compensation ~ '^[£$]?[0-9]+'`,
       sql`SELECT COUNT(*) as count FROM jobs WHERE is_active = true AND role_category = 'Executive' AND (is_remote = true OR workplace_type = 'Remote')`
     ])
     return {
